+++ src/dns_exporter/collector.py
@@ -29,7 +29,6 @@ from dns_exporter.metrics import (
     dnsexp_dns_queries_total,
     dnsexp_dns_responsetime_seconds,
     dnsexp_scrape_failures_total,
-    get_dns_failure_metric,
     get_dns_qtime_metric,
     get_dns_success_metric,
     get_dns_ttl_metric,
@@ -81,7 +80,6 @@ class DNSCollector(Collector):
         """Describe the metrics that are to be returned by this collector."""
         yield get_dns_qtime_metric()
         yield get_dns_success_metric()
-        yield get_dns_failure_metric()
         yield get_dns_ttl_metric()
         yield from self.collect_up()
 
@@ -128,11 +126,11 @@ class DNSCollector(Collector):
         except dns.exception.Timeout:
             # configured timeout was reached before a response arrived
             reason = "timeout"
-            yield from self.yield_failure_reason_metric(failure_reason=reason, config=self.config)
+            self.increase_failure_reason_metric(failure_reason=reason, config=self.config)
         except ConnectionRefusedError:
             # server actively refused the connection
             reason = "connection_error"
-            yield from self.yield_failure_reason_metric(
+            self.increase_failure_reason_metric(
                 failure_reason=reason,
                 config=self.config,
             )
@@ -140,7 +138,7 @@ class DNSCollector(Collector):
             # raised by multiple protocols on ICMP unreach
             logger.debug(f"Protocol {self.config.protocol} got OSError '{e}', exception follows", exc_info=True)
             reason = "connection_error"
-            yield from self.yield_failure_reason_metric(
+            self.increase_failure_reason_metric(
                 failure_reason=reason,
                 config=self.config,
             )
@@ -151,7 +149,7 @@ class DNSCollector(Collector):
                 exc_info=True,
             )
             reason = str(e)
-            yield from self.yield_failure_reason_metric(failure_reason=reason, config=self.config)
+            self.increase_failure_reason_metric(failure_reason=reason, config=self.config)
         except Exception:  # noqa: BLE001
             logger.warning(
                 f"""Caught an unknown exception while looking up qname {self.config.query_name} using server
@@ -160,7 +158,7 @@ class DNSCollector(Collector):
                 exc_info=True,
             )
             reason = "other_failure"
-            yield from self.yield_failure_reason_metric(failure_reason=reason, config=self.config)
+            self.increase_failure_reason_metric(failure_reason=reason, config=self.config)
 
         # clock it
         qtime = time.time() - start
@@ -215,11 +213,11 @@ class DNSCollector(Collector):
         logger.debug("Validating response and yielding remaining metrics")
         try:
             self.validate_response(response=response)
-            yield from self.yield_failure_reason_metric(failure_reason="")
+            self.increase_failure_reason_metric(failure_reason="")
             yield get_dns_success_metric(1)
         except ValidationError as E:
             logger.exception(f"Validation failed: {E.args[1]}")
-            yield from self.yield_failure_reason_metric(failure_reason=E.args[1], config=self.config)
+            self.increase_failure_reason_metric(failure_reason=E.args[1], config=self.config)
             yield get_dns_success_metric(0)
 
     def handle_response_options(self, response: Message) -> None:
@@ -616,10 +614,10 @@ class DNSCollector(Collector):
         self.validate_response_rrs(response=response)
 
     @staticmethod
-    def yield_failure_reason_metric(
+    def increase_failure_reason_metric(
         failure_reason: str,
         config: Config | None = None,
-    ) -> Iterator[CounterMetricFamily]:
+    ) -> None:
         """This method is used to maintain failure metrics.
 
         If an empty string is passed as failure_reason (meaning success) the failure counters will not be incremented.
@@ -633,24 +631,17 @@ class DNSCollector(Collector):
             proxy = "none"
 
         # was there a failure?
-        if failure_reason:
-            # is it a valid failure reason?
-            if failure_reason not in FAILURE_REASONS:
-                # unknown failure_reason, this is a bug
-                raise UnknownFailureReasonError(failure_reason)
-            # increase the global failure counter
-            dnsexp_scrape_failures_total.labels(reason=failure_reason, server=server, proxy=proxy).inc()
-        # get the failure metric
-        fail = get_dns_failure_metric()
-        # initialise all labels in the per-scrape metric,
-        # loop over known failure reasons
-        for reason in FAILURE_REASONS:
-            # set counter to 1 on match (custom collector - the metrics only exist during the scrape)
-            if reason == failure_reason:
-                fail.add_metric([reason, server, proxy], 1)
-            else:
-                fail.add_metric([reason, server, proxy], 0)
-        yield fail
+        if not failure_reason:
+            return
+
+        # is it a valid failure reason?
+        if failure_reason not in FAILURE_REASONS:
+            # unknown failure_reason, this is a bug
+            raise UnknownFailureReasonError(failure_reason)
+
+        # increase the global failure counter
+        dnsexp_scrape_failures_total.labels(reason=failure_reason, server=server, proxy=proxy).inc()
+        return
 
 
 class FailCollector(DNSCollector):
@@ -668,4 +659,4 @@ class FailCollector(DNSCollector):
         yield get_dns_qtime_metric()
         yield get_dns_ttl_metric()
         yield get_dns_success_metric(value=0)
-        yield from self.yield_failure_reason_metric(failure_reason=self.reason)
+        self.increase_failure_reason_metric(failure_reason=self.reason)