+++ src/tests/test_examples.py
@@ -28,9 +28,9 @@ def test_list_of_servers(prometheus_server, dns_exporter_param_config):
     """Test the list_of_servers snippets from the docs."""
     for _ in range(15):
         r = requests.get(
-            'http://127.0.0.1:9092/api/v1/query?query=sum(dnsexp_failures_total{job="dnsexp_doh_gmail_mx"})',
+            "http://127.0.0.1:9092/api/v1/query?query=dnsexp_dns_query_success==1",
         )
-        if len(r.json()["data"]["result"]) > 0 and r.json()["data"]["result"][0]["value"][1] == "0":
+        if len(r.json()["data"]["result"]) > 0 and r.json()["data"]["result"][0]["value"][1] == "1":
             break
         time.sleep(1)
     else:
@@ -51,9 +51,9 @@ def test_list_of_names(caplog, prometheus_server, dns_exporter_param_config):
     """Test the list_of_names snippets from the docs."""
     for _ in range(15):
         r = requests.get(
-            'http://127.0.0.1:9092/api/v1/query?query=sum(dnsexp_failures_total{job="dnsexp_quad9_mx"})',
+            "http://127.0.0.1:9092/api/v1/query?query=dnsexp_dns_query_success==1",
         )
-        if len(r.json()["data"]["result"]) > 0 and r.json()["data"]["result"][0]["value"][1] == "0":
+        if len(r.json()["data"]["result"]) > 0 and r.json()["data"]["result"][0]["value"][1] == "1":
             break
         time.sleep(1)
     else: