+++ src/tests/test_exporter.py
@@ -90,7 +90,7 @@ def test_invalid_qs_ip(dns_exporter_example_config):
             "ip": "notanip",
         },
     )
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_ip",server="none"} 1.0' in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_invalid_configfile_ip(caplog, exporter):
@@ -108,7 +108,7 @@ def test_missing_query_name(dns_exporter_example_config):
             "server": "dns.google",
         },
     )
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_query_name",server="none"} 1.0' in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_missing_server(dns_exporter_example_config):
@@ -119,7 +119,7 @@ def test_missing_server(dns_exporter_example_config):
             "query_name": "example.com",
         },
     )
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_server",server="none"} 1.0' in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_undefined_module(dns_exporter_example_config, caplog):
@@ -130,7 +130,7 @@ def test_undefined_module(dns_exporter_example_config, caplog):
             "module": "notamodule",
         },
     )
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_module",server="none"} 1.0' in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_unknown_config_key(dns_exporter_example_config, caplog):
@@ -142,7 +142,7 @@ def test_unknown_config_key(dns_exporter_example_config, caplog):
             "foo": "bar",
         },
     )
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_config",server="none"} 1.0' in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_ip_family_conflict(dns_exporter_example_config, caplog):
@@ -156,7 +156,7 @@ def test_ip_family_conflict(dns_exporter_example_config, caplog):
             "ip": "192.0.2.53",
         },
     )
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_ip",server="none"} 1.0' in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_ip_conflict(dns_exporter_example_config, caplog):
@@ -170,14 +170,14 @@ def test_ip_conflict(dns_exporter_example_config, caplog):
             "family": "ipv4",
         },
     )
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_ip",server="none"} 1.0' in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_ip_and_hostname(dns_exporter_example_config, caplog):
     """Make sure using an IP and hostname results in the IP overriding the server hostname DNS lookup."""
     caplog.clear()
     caplog.set_level(logging.DEBUG)
-    requests.get(
+    r = requests.get(
         "http://127.0.0.1:25353/query",
         params={
             "server": "dns.google",
@@ -187,6 +187,7 @@ def test_ip_and_hostname(dns_exporter_example_config, caplog):
         },
     )
     assert "Using server IP 8.8.4.4 (from config) for the DNS server connection" in caplog.text
+    assert "dnsexp_dns_query_success 1.0" in r.text
 
 
 def test_unresolvable_server(dns_exporter_example_config, caplog):
@@ -200,7 +201,7 @@ def test_unresolvable_server(dns_exporter_example_config, caplog):
             "query_name": "example.com",
         },
     )
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_server",server="none"} 1.0' in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_ipv6_family(dns_exporter_example_config, caplog):
@@ -231,7 +232,7 @@ def test_ipv7_family(dns_exporter_example_config, caplog):
             "family": "ipv7",
         },
     )
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_family",server="none"} 1.0' in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 # run this test last
@@ -320,6 +321,7 @@ def test_tcp(dns_exporter_example_config, caplog):
     )
     assert 'transport="TCP"' in r.text
     assert "Protocol tcp got a DNS query response over TCP" in caplog.text
+    assert "dnsexp_dns_query_success 1.0" in r.text
 
 
 def test_udptcp(dns_exporter_example_config, caplog):
@@ -337,6 +339,7 @@ def test_udptcp(dns_exporter_example_config, caplog):
     )
     assert 'transport="UDP"' in r.text
     assert "Protocol udptcp got a DNS query response over UDP" in caplog.text
+    assert "dnsexp_dns_query_success 1.0" in r.text
 
 
 def test_dot(dns_exporter_example_config, caplog):
@@ -354,6 +357,7 @@ def test_dot(dns_exporter_example_config, caplog):
     )
     assert 'transport="TCP"' in r.text
     assert "Protocol dot got a DNS query response over TCP" in caplog.text
+    assert "dnsexp_dns_query_success 1.0" in r.text
 
 
 def test_doh(dns_exporter_example_config, caplog):
@@ -372,6 +376,7 @@ def test_doh(dns_exporter_example_config, caplog):
     assert 'transport="TCP"' in r.text
     assert 'protocol="doh"' in r.text
     assert "Protocol doh got a DNS query response over TCP" in caplog.text
+    assert "dnsexp_dns_query_success 1.0" in r.text
 
 
 def test_doq(dns_exporter_example_config, caplog, recwarn):
@@ -390,6 +395,7 @@ def test_doq(dns_exporter_example_config, caplog, recwarn):
     assert 'transport="QUIC"' in r.text
     assert 'protocol="doq"' in r.text
     assert "Protocol doq got a DNS query response over QUIC" in caplog.text
+    assert "dnsexp_dns_query_success 1.0" in r.text
 
 
 def test_validate_rcode(dns_exporter_example_config, caplog):
@@ -402,10 +408,8 @@ def test_validate_rcode(dns_exporter_example_config, caplog):
             "family": "ipv4",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_rcode",server="udp://dns.google:53"} 1.0' in r.text
-    )
     assert 'rcode="NXDOMAIN"' in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_validate_flags_fail_if_any_absent(dns_exporter_example_config, caplog):
@@ -419,9 +423,7 @@ def test_validate_flags_fail_if_any_absent(dns_exporter_example_config, caplog):
             "module": "has_ad",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_flags",server="udp://dns.google:53"} 1.0' in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_validate_flags_fail_if_any_present(dns_exporter_example_config, caplog):
@@ -435,9 +437,7 @@ def test_validate_flags_fail_if_any_present(dns_exporter_example_config, caplog)
             "module": "has_no_ad",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_flags",server="udp://dns.google:53"} 1.0' in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_validate_flags_fail_if_all_present(dns_exporter_example_config, caplog):
@@ -451,9 +451,7 @@ def test_validate_flags_fail_if_all_present(dns_exporter_example_config, caplog)
             "module": "fail_not_auth",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_flags",server="udp://dns.google:53"} 1.0' in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_validate_flags_fail_if_all_absent(dns_exporter_example_config, caplog):
@@ -467,9 +465,7 @@ def test_validate_flags_fail_if_all_absent(dns_exporter_example_config, caplog):
             "module": "fail_recursive",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_flags",server="udp://dns.google:53"} 1.0' in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_validate_flags_fail_if_all_present_2(dns_exporter_example_config, caplog):
@@ -483,10 +479,7 @@ def test_validate_flags_fail_if_all_present_2(dns_exporter_example_config, caplo
             "module": "fail_not_auth",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_flags",server="udp://dns.quad9.net:53"} 1.0'
-        in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_validate_flags_fail_if_all_absent_2(dns_exporter_example_config, caplog):
@@ -500,9 +493,7 @@ def test_validate_flags_fail_if_all_absent_2(dns_exporter_example_config, caplog
             "module": "fail_recursive",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_flags",server="udp://dns.google:53"} 1.0' in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_validate_flags_fail_if_all_present_3(dns_exporter_example_config, caplog):
@@ -546,10 +537,7 @@ def test_validate_rr_fail_if_matches_regexp(dns_exporter_example_config, caplog)
             "module": "fail_auth_k_root",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_answer_rrs",server="udp://l.root-servers.net:53"} 1.0'
-        in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_validate_rrs_fail_if_all_match_regexp(dns_exporter_example_config, caplog):
@@ -564,10 +552,7 @@ def test_validate_rrs_fail_if_all_match_regexp(dns_exporter_example_config, capl
             "module": "fail_additional_root",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_additional_rrs",server="udp://k.root-servers.net:53"} 1.0'
-        in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_validate_rrs_fail_if_all_match_regexp_2(dns_exporter_example_config, caplog):
@@ -597,10 +582,7 @@ def test_validate_rrs_fail_if_not_matches_regexp(dns_exporter_example_config, ca
             "module": "fail_answer_root",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_answer_rrs",server="udp://dns.google:53"} 1.0'
-        in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_validate_rrs_fail_if_none_matches_regexp(dns_exporter_example_config, caplog):
@@ -615,10 +597,7 @@ def test_validate_rrs_fail_if_none_matches_regexp(dns_exporter_example_config, c
             "module": "fail_answer_root_none",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_answer_rrs",server="udp://dns.google:53"} 1.0'
-        in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_validate_rrs_fail_if_none_matches_regexp_2(
@@ -636,10 +615,7 @@ def test_validate_rrs_fail_if_none_matches_regexp_2(
             "module": "fail_answer_root_none",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="invalid_response_answer_rrs",server="udp://dns.google:53"} 1.0'
-        in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_edns_pad(dns_exporter_example_config, caplog):
@@ -729,7 +705,7 @@ def test_invalid_integer(dns_exporter_example_config, caplog):
     )
     assert "Unable to validate integer for key edns_bufsize" in caplog.text
     assert "ValueError: invalid literal for int() with base 10: 'foo'" in caplog.text
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_config",server="none"} 1.0' in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_configure_rrvalidator(caplog, exporter):
@@ -788,10 +764,7 @@ def test_catch_unknown_exception(
             "protocol": "doh",
         },
     )
-    assert (
-        'dnsexp_failures_total{proxy="none",reason="other_failure",server="doh://dns.google:443/dns-query"} 1.0'
-        in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 ### ttl tests
@@ -808,6 +781,7 @@ def test_ttl(dns_exporter_example_config, caplog):
         },
     )
     assert 'rr_value="91.239.100.100"' in r.text
+    assert "dnsexp_dns_query_success 1.0" in r.text
 
 
 def test_no_ttl(dns_exporter_no_main_no_config, caplog):
@@ -824,6 +798,7 @@ def test_no_ttl(dns_exporter_no_main_no_config, caplog):
         },
     )
     assert 'rr_value="91.239.100.100"' not in r.text
+    assert "dnsexp_dns_query_success 1.0" in r.text
 
 
 ### connection refused
@@ -843,15 +818,7 @@ def test_connection_error_server(dns_exporter_example_config, caplog, protocol):
             "timeout": 1,
         },
     )
-    # this is handled a bit differently depending on the ICMP error (if any) received from the network
-    server = f"{protocol}://192.0.2.42:420"
-    if protocol == "doh":
-        server = f"{server}/dns-query"
-    if (
-        'dnsexp_failures_total{proxy="none",reason="connection_error",server="%s"} 1.0' % server not in r.text
-        and 'dnsexp_failures_total{proxy="none",reason="timeout",server="%s"} 1.0' % server not in r.text
-    ):
-        raise AssertionError(protocol)
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 @pytest.mark.parametrize("protocol", ["udp", "tcp", "udptcp", "dot", "doh", "doq"])
@@ -868,14 +835,7 @@ def test_timeout_server(dns_exporter_example_config, caplog, protocol):
             "timeout": 0.00001,
         },
     )
-    server = f"{protocol}://192.0.2.42:420"
-    if protocol == "doh":
-        server = f"{server}/dns-query"
-    if protocol == "doh":
-        # doh raises httpx.ConnectError
-        assert 'dnsexp_failures_total{proxy="none",reason="connection_error",server="%s"} 1.0' % server in r.text
-    else:
-        assert 'dnsexp_failures_total{proxy="none",reason="timeout",server="%s"} 1.0' % server in r.text
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 def test_collect_ttl_value_length(dns_exporter_example_config, caplog):
@@ -892,3 +852,4 @@ def test_collect_ttl_value_length(dns_exporter_example_config, caplog):
         },
     )
     assert 'rr_value="8.8"' in r.text
+    assert "dnsexp_dns_query_success 1.0" in r.text