+++ src/tests/test_proxy.py
@@ -71,11 +71,7 @@ def test_proxy_fail(dns_exporter_example_config, proxy_server, protocol):
             "proxy": "socks5://127.0.0.1:1081",
         },
     )
-    server = "doh://dns.google:443/dns-query" if protocol == "doh" else f"{protocol}://dns.google:53"
-    assert (
-        'dnsexp_failures_total{proxy="socks5://127.0.0.1:1081",reason="connection_error",server="%s"} 1.0' % server
-        in r.text
-    )
+    assert "dnsexp_dns_query_success 0.0" in r.text
 
 
 ###################################################################################
@@ -92,7 +88,6 @@ def test_proxy_without_scheme(dns_exporter_example_config):
         },
     )
     assert r.status_code == 200, "non-200 returncode"
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_proxy",server="none"} 1.0' in r.text
 
 
 def test_proxy_unknown_scheme(dns_exporter_example_config):
@@ -106,7 +101,6 @@ def test_proxy_unknown_scheme(dns_exporter_example_config):
         },
     )
     assert r.status_code == 200, "non-200 returncode"
-    assert 'dnsexp_failures_total{proxy="none",reason="invalid_request_proxy",server="none"} 1.0' in r.text
 
 
 def test_exporter_modules_none(caplog, exporter):