+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Minimum supported DNSPython version is now 2.5.0
 - Log at level warning when encountering an unknown failure
 - Remove failure reason `connection_refused` and report it as `connection_error` along with the rest of socket related errors.
+- Remove `dnsexp_failures_total` from per-scrape metrics. Failure reasons are now tracked only under `/metrics`.
 
 
 ## [v1.0.0-beta6] - 2024-03-01